+++ src/id/index.ts
@@ -1,2 +1,0 @@
-export * from "./unique-id-for.js";
-export * from "./unique-id.js";