+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 20.0
 + **Breaking:** Remove `IndexFor`. Use `Index` instead.
++ **Breaking:** Remove `Range` from `rvx/dom`.
 
 ## 19.7
 + **Deprecated:** `created` will be removed in the next major version. `created` is now an alias for `useMicrotask` and has lifecycle support.