+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 20.0
 + **Breaking:** Remove `IndexFor`. Use `Index` instead.
 + **Breaking:** Remove `Range` from `rvx/dom`.
++ **Breaking:** Remove `created`. Use `useMicrotask` instead.
++ **Breaking:** Move `string` & `optionalString` to `rvx/convert`.
 
 ## 19.7
 + **Deprecated:** `created` will be removed in the next major version. `created` is now an alias for `useMicrotask` and has lifecycle support.